Widen FontSize from int to float end-to-end (#4316)
* Widen FontSize from int to float end-to-end (#4304)

TextRuntime.FontSize, BmfcSave.FontSize/GetFontCacheFileNameFor, SkiaGum/RaylibGum's Text.FontSize, and the shared inline-BBCode font-size stack now all take float, letting a fractional size (e.g. 18.5) render distinctly from its neighboring integers.

KernSmith rasterizes the fractional value natively (FontGeneratorOptions.Size is float on the FreeType backend Gum registers). The legacy bmfont.exe/GDI backend can't, so BmfcSave.Save() and the bmfont.exe size-estimation heuristic round to the nearest whole pixel via the new BmfcSave.GetGdiRoundedFontSize.

Tool property grid is untouched; this ships as a runtime-only API.

Closes #4304


* Add fractional FontSize column to FontPlayground sample for manual testing

Issue #4304: 20 Text instances stacked vertically, FontSize starting at 10 and increasing by 0.25 per row, so KernSmith's fractional rasterization can be inspected visually.

* Replace fractional FontSize column with a single scroll-driven text

Removes the 20-row strip (its 0.25 fixed steps made it hard to see continuous zoom-driven behavior) in favor of one Text whose FontSize tracks the mouse wheel directly and displays 'Font size: XX.XX'. Deliberately NOT tied to camera.Zoom -- this text stays in unzoomed screen space so KernSmith's freshly-baked raster always matches its own on-screen size 1:1, with no compounding from the camera transform.

* Position the fractional FontSize demo text first, not last

Y=560 (topmost of the zoom-demo group) instead of Y=660, and built before the existing oversampling pair -- it was tacked onto the very end at the bottom of the window.

* Reserve a real top band for the fractional FontSize demo text

Y=560 was 'first' only within a Game1.cs-local group, still well below the controls panel and preview texts -- not actually first on the page. Now the fractional-FontSize text sits at Y=4 in a dedicated top band, with controlsPanel/preview texts (FontPlaygroundScreen, shared) and the zoom-demo pair (Game1.cs) all shifted down to stay clear of it. FractionalFontSizeMax capped at 48 (was 96) so the text can never grow large enough to run into the controls panel below; window height bumped to 820 to fit the shift.

* Fix compounding camera zoom and non-reversible zoom formula in FontPlayground

Two real bugs:

1. The fractional-FontSize text rendered through the same camera the mouse wheel also zooms, so its own FontSize growth compounded with the camera's magnification on top -- pixelation, drift from the zoom pivot, and size growing much faster than the displayed number implied. Moved it onto a dedicated screen-space layer (LayerCameraSettings.IsInScreenSpace), which the main camera ignores entirely, so its on-screen size/position now come only from its own FontSize/X/Y.

2. camera.Zoom *= (1 + delta*k) for scroll-up and *= (1 - delta*k) for scroll-down are not inverses -- (1+k)(1-k) = 1-k^2 < 1, so equal-and-opposite scrolling loses a little zoom every round trip. Replaced with an exponential step (Pow(1.001, delta)), which is its own exact inverse. Applied to both camera.Zoom and the fractional FontSize update, since I'd copied the same broken formula into the new code.

* Note that automatic crisp zoom is tracked separately in #4317

Clarifies scope: this demo shows fractional FontSize triggering fresh KernSmith rasterization, not camera-zoom-driven crisp text (which needs a FontScale-free compensation mechanism that doesn't exist yet).

---------

+++ Gum/Wireframe/CustomSetPropertyOnRenderable.cs
@@ -1431,7 +1431,7 @@ public partial class CustomSetPropertyOnRenderable
     // The eight font-resolution stacks (mirroring the MonoGame path). Inline [FontSize]/[IsBold]/etc. tags
     // push their value on open and pop on close, so a run resolves to the font implied by every tag currently
     // open over it - the whole reason nested markup needs a stack rather than a flat lookup.
-    static Stack<int> fontSizeStack = new Stack<int>();
+    static Stack<float> fontSizeStack = new Stack<float>();
     static Stack<string> fontNameStack = new Stack<string>();
     static Stack<int> outlineThicknessStack = new Stack<int>();
     static Stack<bool> useFontSmoothingStack = new Stack<bool>();
@@ -1476,7 +1476,7 @@ public partial class CustomSetPropertyOnRenderable
     /// the caller.
     /// </summary>
     static BmfcSave BuildInlineRunBmfcSave(
-        int fontSize, int outlineThickness, bool useFontSmoothing, bool isItalic, bool isBold,
+        float fontSize, int outlineThickness, bool useFontSmoothing, bool isItalic, bool isBold,
         string? fontFilePath, string fontName)
     {
         BmfcSave bmfcSave = new BmfcSave();
@@ -1888,7 +1888,7 @@ public partial class CustomSetPropertyOnRenderable
                     castedValue = GetAndCreateFontIfNecessary();
                     break;
                 case "FontSize":
-                    if (int.TryParse(tag.Argument, NumberStyles.Integer, CultureInfo.InvariantCulture, out int parsedSize))
+                    if (float.TryParse(tag.Argument, NumberStyles.Float, CultureInfo.InvariantCulture, out float parsedSize))
                     {
                         fontSizeStack.Push(parsedSize);
                         castedValue = GetAndCreateFontIfNecessary();
@@ -1897,7 +1897,7 @@ public partial class CustomSetPropertyOnRenderable
                         // rasterized at the requested size, so store the raw pixel size as a float and let the
                         // renderer scale the base atlas to it (a blurry-but-correct approximation). The XNA
                         // path never needs this - it can new BitmapFont(...) at any size.
-                        castedValue ??= (float)parsedSize;
+                        castedValue ??= parsedSize;
 #endif
                     }
                     else
